requirements.txt
google.api_core==1.20.0
google-cloud-storage==1.28.0
apache-beam[gcp]==2.21.0
setuptools>=40.3.0
google-cloud-pubsub==1.6.0
